requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.4
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Flask-WTF==1.1.1
SQLAlchemy==2.0.6
Werkzeug==2.3.7
Jinja2==3.1.4
itsdangerous==2.1.2
MarkupSafe==2.1.2
alembic==1.10.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
WTForms==3.0.1
email_validator==2.1.0.post1
